+++ stdlib/LinearAlgebra/src/deprecated.jl
@@ -1356,3 +1356,7 @@ export eigfact!
 export schurfact!
 @deprecate(schurfact!(A::StridedMatrix{T}, B::StridedMatrix{T}) where {T<:BlasFloat}, schur!(A, B))
 @deprecate(schurfact!(A::StridedMatrix{<:BlasFloat}), schur!(A))
+
+# deprecate lqfact! to lq!
+export lqfact!
+@deprecate(lqfact!(A::StridedMatrix{<:BlasFloat}), lq!(A))